requirements.txt
selenium>=4.0.0
webdriver-manager>=3.8.0
PyQt5>=5.15.0
pyautogui>=0.9.53
pytesseract>=0.3.10
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
flask>=2.0.0
gitpython>=3.1.0
keyboard>=0.13.5
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
